downstream: refactoring code to remove listener hard deps
