+++ source/extensions/matching/actions/format_string/config.h
@@ -7,7 +7,6 @@
 #include "envoy/server/factory_context.h"
 
 #include "source/common/matcher/matcher.h"
-#include "source/server/filter_chain_manager_impl.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -16,22 +15,24 @@ namespace Actions {
 namespace FormatString {
 
 class ActionImpl : public Matcher::ActionBase<envoy::config::core::v3::SubstitutionFormatString,
-                                              Server::FilterChainBaseAction> {
+                                              Server::Configuration::FilterChainBaseAction> {
 public:
   ActionImpl(const Formatter::FormatterConstSharedPtr& formatter) : formatter_(formatter) {}
-  const Network::FilterChain* get(const Server::FilterChainsByName& filter_chains_by_name,
-                                  const StreamInfo::StreamInfo& info) const override;
+  const Network::FilterChain*
+  get(const Server::Configuration::FilterChainsByName& filter_chains_by_name,
+      const StreamInfo::StreamInfo& info) const override;
 
 private:
   const Formatter::FormatterConstSharedPtr formatter_;
 };
 
-class ActionFactory : public Matcher::ActionFactory<Server::FilterChainActionFactoryContext> {
+using FilterChainActionFactoryContext = Server::Configuration::ServerFactoryContext;
+class ActionFactory : public Matcher::ActionFactory<FilterChainActionFactoryContext> {
 public:
   std::string name() const override { return "envoy.matching.actions.format_string"; }
   Matcher::ActionFactoryCb
   createActionFactoryCb(const Protobuf::Message& proto_config,
-                        Server::FilterChainActionFactoryContext& context,
+                        FilterChainActionFactoryContext& context,
                         ProtobufMessage::ValidationVisitor& validator) override;
   ProtobufTypes::MessagePtr createEmptyConfigProto() override {
     return std::make_unique<envoy::config::core::v3::SubstitutionFormatString>();
