+++ source/server/BUILD
@@ -101,11 +101,22 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "listener_stats",
+    hdrs = [
+        "listener_stats.h",
+    ],
+    deps = [
+        "//source/common/stats:timespan_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "active_listener_base",
     hdrs = [
         "active_listener_base.h",
     ],
     deps = [
+        ":listener_stats",
         "//envoy/network:connection_handler_interface",
         "//envoy/network:listener_interface",
         "//envoy/stats:timespan_interface",
