+++ source/server/filter_chain_manager_impl.h
@@ -103,12 +103,6 @@ private:
 using FilterChainActionFactoryContext = Configuration::ServerFactoryContext;
 using FilterChainsByName = absl::flat_hash_map<std::string, Network::DrainableFilterChainSharedPtr>;
 
-class FilterChainBaseAction : public Matcher::Action {
-public:
-  virtual const Network::FilterChain* get(const FilterChainsByName& filter_chains_by_name,
-                                          const StreamInfo::StreamInfo& info) const PURE;
-};
-
 class FilterChainImpl : public Network::DrainableFilterChain {
 public:
   FilterChainImpl(Network::DownstreamTransportSocketFactoryPtr&& transport_socket_factory,
